feat: adjust loadApp method
