--- packages/runtime/core/src/utils/eventTypes.ts
@@ -1,17 +1,0 @@
-export const BEFORE_BOOTSTRAP = 'beforeBootstrap'; // 启动之前
-export const BOOTSTRAP = 'bootstrap'; // 主应用启动
-export const REGISTER_APP = 'registerApp'; // 注册 app
-export const BEFORE_LOAD_APP = 'beforeLoadApp'; // 加载 app 之前
-export const START_LOAD_APP = 'startLoadApp'; // 开始下载子应用
-export const CREATE_APP = 'createApp'; // 创建一个 app
-export const LOAD_APP_ERROR = 'loadAppError'; // 加载 app 发生错误
-export const END_LOAD_APP = 'endLoadApp'; // 结束下载子应用
-export const START_COMPILE_APP = 'startCompileApp'; // 开始编译脚本
-export const END_COMPILE_APP = 'endCompileApp'; // 结束编译脚本
-export const ERROR_COMPILE_APP = 'errorCompileApp'; // 编译脚本失败
-export const START_MOUNT_APP = 'startMountApp'; // 开始挂在子应用
-export const END_MOUNT_APP = 'endMountApp'; // 结束挂载子应用
-export const START_UNMOUNT_APP = 'startUnmountApp'; // 开始卸载子应用
-export const END_UNMOUNT_APP = 'endUnmountApp'; // 结束卸载子应用
-export const ERROR_MOUNT_APP = 'errorMountApp'; // 挂载子应用异常
-export const ERROR_UNMOUNT_APP = 'errorUnmountApp'; // 卸载子应用异常
